starting to work on streaming cleaning up javascript code
