+++ nodes/config/OllamaStructureConfig.py
@@ -45,6 +45,7 @@ class gtUIOllamaStructureConfig(gtUIBaseConfig):
         temperature = kwargs.get("temperature", 0.7)
         base_url = kwargs.get("base_url", ollama_base_url)
         port = kwargs.get("port", ollama_port)
+        stream = kwargs.get("stream", False)
 
         image_generation_driver = kwargs.get(
             "image_generation_driver", DummyImageGenerationDriver()
@@ -57,6 +58,7 @@ class gtUIOllamaStructureConfig(gtUIBaseConfig):
                 temperature=temperature,
                 host=f"{base_url}:{port}",
                 max_attempts=max_attempts,
+                stream=stream,
             ),
             embedding_driver=DummyEmbeddingDriver(),
             image_generation_driver=image_generation_driver,
